Ensure the status and status date fields are empty when the page loads. Populate these fields with output parameters when the recipe is saved.

+++ Database Files/sprocs/CookbookUpdate.sql
@@ -2,7 +2,7 @@ CREATE OR ALTER PROC dbo.CookbookUpdate(
     @CookbookID int OUTPUT, -- Corrected parameter name
     @UsersID int,
     @CookbookName varchar(100),
-    @Price int,
+    @Price DECIMAL,
     @DateCreated DATETIME,
     @Active int
 )
